requirements.txt
appdirs==1.4.2
MySQL-python==1.2.5
packaging==16.8
pyparsing==2.1.10
python-firebase==1.2
requests==1.1.0
six==1.10.0
